requirements.txt

Django==4.1.1
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
googlemaps==4.6.0
gunicorn==20.1.0
permission==0.4.1
Pillow==9.2.0
psycopg2==2.9.3
PyJWT==2.5.0
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.3
tzdata==2022.4

